# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * Returns all configured workflow tools
 * @summary List all tools
 */
export const ListToolsResponseItem = zod.object({
  "id": zod.string(),
  "title": zod.string(),
  "description": zod.string(),
  "icon": zod.string(),
  "url": zod.string()
})
export const ListToolsResponse = zod.array(ListToolsResponseItem)


/**
 * Adds a new workflow tool
 * @summary Add a tool
 */
export const CreateToolBody = zod.object({
  "title": zod.string(),
  "description": zod.string(),
  "icon": zod.string(),
  "url": zod.string()
})


/**
 * Updates an existing workflow tool
 * @summary Update a tool
 */
export const UpdateToolParams = zod.object({
  "id": zod.coerce.string()
})

export const UpdateToolBody = zod.object({
  "title": zod.string(),
  "description": zod.string(),
  "icon": zod.string(),
  "url": zod.string()
})

export const UpdateToolResponse = zod.object({
  "id": zod.string(),
  "title": zod.string(),
  "description": zod.string(),
  "icon": zod.string(),
  "url": zod.string()
})


/**
 * Removes a workflow tool
 * @summary Delete a tool
 */
export const DeleteToolParams = zod.object({
  "id": zod.coerce.string()
})