# lib/api-zod/src/generated/types/createToolBody.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateToolBody {
  title: string;
  description: string;
  icon: string;
  url: string;
}