# lib/api-zod/src/generated/types/tool.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Tool {
  id: string;
  title: string;
  description: string;
  icon: string;
  url: string;
}